% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dichoplot_table.R
\name{create_dichoplot_table}
\alias{create_dichoplot_table}
\title{Create a dichotomized table of a selection of variables including the national mean}
\usage{
create_dichoplot_table(data, vars, restrict = TRUE, restrict_level = 10L)
}
\arguments{
\item{data}{a C or E SHURP-dataframe with dichotomized items}

\item{vars}{a vector passing variable names}

\item{restrict}{defaults to TRUE, passed to shurp2018::prop.yes()}

\item{restrict_level}{defaults to TRUE, passed to shurp2018::prop.yes()}
}
\value{

}
\description{
Create a dichotomized table of a selection of variables including the national mean
}
\examples{
\dontrun{create_dichoplot_table(shurp2018_C_NA, vars = "C1dich")}
}
