% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dichotomization_funcs.R
\name{dichotomize.C}
\alias{dichotomize.C}
\title{Automatically dichotomoize the C dataframes of SHURP2018}
\usage{
dichotomize.C(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{

}
\description{
Automatically dichotomoize the C dataframes of SHURP2018
}
