% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dichotomization_funcs.R
\name{dichotomize.E}
\alias{dichotomize.E}
\title{Automatically dichotomoize the E dataframes of SHURP2018}
\usage{
dichotomize.E(df)
}
\arguments{
\item{df}{a dataframe

@section How to?
Simply assign your dataframe like this:
dichotomized_data <- \code{dichotomize.E(your_data)}}
}
\description{
Automatically dichotomoize the E dataframes of SHURP2018
}
