% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggBernchmarketing.R
\name{ggBenchmarketing}
\alias{ggBenchmarketing}
\title{Create the so-called "benchmarketing chart" for overview of percentage agreement}
\usage{
ggBenchmarketing(
  df,
  col,
  title = NULL,
  text_caption = TRUE,
  labs = TRUE,
  language = "german"
)
}
\arguments{
\item{df}{a dataframe created by create_dichoplot_table()}

\item{col}{a variable holding pct-agreement}

\item{title}{either a string or a column from label$data}

\item{text_caption}{TRUE/FALSE, default TRUE, if caption holding the corresponding}

\item{labs}{TRUE/FALSE, default TRUE, if Title/subtitle should be displayed}

\item{language}{"german" or "french" - default "german", select subtitle and}
}
\value{

}
\description{
Create the so-called "benchmarketing chart" for overview of percentage agreement
}
