% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{label_data}
\alias{label_data}
\title{Item labels used for params reporting}
\format{
A data frame with 2 rows and 91 variables:
\describe{
\item{var_name}{indicating what label}
\item{german}{german text to label}
\item{french}{french text to label}
}
}
\source{
SHURP 2018 study
}
\usage{
label_data
}
\description{
A dataset containing item labels
}
\keyword{datasets}
