% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_funcs.R
\name{n.valid}
\alias{n.valid}
\title{get sum of not NAs of a vector}
\usage{
n.valid(x)
}
\arguments{
\item{x}{a vector}
}
\value{
\code{n.valid()} returns the number of fields not being NA in a vector
}
\description{
get sum of not NAs of a vector
}
