% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_funcs.R
\name{n.yes}
\alias{n.yes}
\title{get number of positive answers in a vector}
\usage{
n.yes(x, cutoff = 1)
}
\arguments{
\item{x}{a vector}

\item{cutoff}{a vector defining positive answers, default = 1 in a binary dataset}
}
\value{
a value
}
\description{
get number of positive answers in a vector
}
