% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_funcs.R
\name{prop.yes}
\alias{prop.yes}
\title{get percentage of positive answers in a vector}
\usage{
prop.yes(x, cutoff = 1, round = T, digits = 2, restrict = F)
}
\arguments{
\item{x}{a vector}

\item{cutoff}{a vector defining positive answers, default = 1 in a binary dataset}

\item{round}{TRUE/FALSE, default = T}

\item{digits}{no. of digits to round, default = 2}

\item{restrict}{should answers be restricted to 'NA' when n < 10 'NA', default F}
}
\description{
get percentage of positive answers in a vector
}
