% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{QC_text}
\alias{QC_text}
\title{QC-Item text data in all languages}
\format{
A data frame with 2 rows and 168 variables:
\describe{
\item{Version}{language, french or german}
\item{C1}{Text to item C1}
...
}
}
\source{
SHURP 2018 study
}
\usage{
QC_text
}
\description{
A dataset containing corresponding item texts for QC (including cutoffs for
}
\keyword{datasets}
