% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{QE_text}
\alias{QE_text}
\title{QE-Item text data in all languages}
\format{
A data frame with 2 rows and 91 variables:
\describe{
\item{Version}{language, french or german}
\item{E1}{Text to item E1}
...
}
}
\source{
SHURP 2018 study
}
\usage{
QE_text
}
\description{
A dataset containing corresponding item texts for QE (including cutoffs for
}
\keyword{datasets}
