% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_report.R
\name{render_report}
\alias{render_report}
\title{Render individual report as PDF - this works only with Rmd-files}
\usage{
render_report(file = "index.Rmd", language, center)
}
\arguments{
\item{file}{the file to render, default = "index.Rmd"}

\item{language}{a character - either "D" or "F"}

\item{center}{a number - choose center number e.g. 999}
}
\value{
creates a PDF output named according to language and center
}
\description{
Render individual report as PDF - this works only with Rmd-files
}
