% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_funcs.R
\name{round_df}
\alias{round_df}
\title{round all numeric variables of a df according to no. of digits}
\usage{
round_df(df, digits)
}
\arguments{
\item{df}{a dataframe}

\item{digits}{a numeric value}
}
\value{
a df with rounded numeric variables
}
\description{
round all numeric variables of a df according to no. of digits
}
