% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SHURP_syn}
\alias{SHURP_syn}
\title{SHURP_syn.}
\format{A data frame with answers from 5311 nurses in  69 variables. More
information is found in the vignette.}
\usage{
SHURP_syn
}
\description{
Synthesized data from the SHURP2013-study.
}
\keyword{datasets}
