% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_funcs.R
\name{getvartext}
\alias{getvartext}
\title{get text of variables from a specified text spreadsheet}
\usage{
getvartext(x)
}
\arguments{
\item{x}{a vector}
}
\value{
the first element of the vector
}
\description{
get text of variables from a specified text spreadsheet
}
\examples{
a <- c(1, 23, 4, 56, 7)
getvartext(a)
}
