% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_funcs.R
\name{getvartext}
\alias{getvartext}
\title{get text of variables from a specified text spreadsheet}
\usage{
getvartext(x, y = 1)
}
\arguments{
\item{x}{a vector}

\item{y}{the nth element of a vector to extract, default = 1 (first)}
}
\value{
returns the nth element of a vector
}
\description{
get text of variables from a specified text spreadsheet
}
\examples{
a <- c(77, 23, 4, 56, 7)
getvartext(a)
}
